[feat] Reimplement weather search feature with new WeatherSearchEvent, offer offline-first saved locations with their weather data

+++ app/src/main/java/com/example/weatherjourney/core/database/model/DailyWeatherEntity.kt
@@ -22,5 +22,5 @@ data class DailyWeatherEntity(
     var maxTemperatures: FloatListHolder,
     var minTemperatures: FloatListHolder,
     @PrimaryKey
-    val locationId: Int
+    val locationId: Long
 )
